refactor: centralize `item` related logic in `item_extractor`
